feat: 수익률 구하는 메소드 구현

+++ src/main/java/lotto/Dollars.java
@@ -26,8 +26,8 @@ public class Dollars {
         return count;
     }
 
-    public int won() {
-        return count * DOLLAR_UNIT;
+    public long won() {
+        return (long) count * DOLLAR_UNIT;
     }
 
     @Override
